% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_dots.R
\name{barplot_dots}
\alias{barplot_dots}
\title{Bar plot with dots
Plot by column having WT and KO next on the same bar and all the dots}
\usage{
barplot_dots(
  table,
  organ.i = NULL,
  stat.i = NULL,
  time.i = NULL,
  marker.i = NULL,
  cell.i = NULL,
  treatment.i = NULL,
  genotype_levels = c("WT", "KO"),
  genotype_labels = genotype_levels,
  identity_bar = "dodge",
  x_value = "cell",
  x_lab = "",
  y_lab = "",
  y_limit = 0,
  title_lab = "",
  x_labels = waiver(),
  y_trans = "identity",
  y_label = waiver(),
  leyend_position = "top",
  x_angle = 0,
  x_hjust = 0.5,
  color_values = colorRamps::primary.colors(),
  shape_values = rep(21, 200),
  fill_values = colorRamps::primary.colors(),
  path_output = NULL,
  w = 10,
  h = 5,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with data comming from the analysis. Columns: genotype, value and experiment (time, mice, treatment, marker, stat, cell)}

\item{organ.i}{optional organ selected to plot (specimen in .fcs file)}

\item{stat.i}{optional statistic selected to plot}

\item{time.i}{optional time selected to plot}

\item{marker.i}{optional marker selected to plot}

\item{cell.i}{optional cell selected to plot}

\item{treatment.i}{optional treatment selected to plot}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{genotype_labels}{name to be display in the legend. In markdown/html format.}

\item{identity_bar}{Position to plot the bar graph. "identity" to pile, "dodge" to put next to each other. Default to dodge. Depending on the value, the points represented are pilled up by strain or not}

\item{x_value}{column name to plot}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{y_limit}{inferior limit for y-axis}

\item{title_lab}{title label}

\item{x_labels}{labels for x axis}

\item{y_trans}{transformation of the y axis}

\item{y_label}{default to waiver. Could be scientific_format()}

\item{leyend_position}{Legend position. Default to top. Could also be right, left or bottom}

\item{x_angle}{Angle to display the x axis}

\item{x_hjust}{Justificacion of the x axis}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{shape_values}{shape to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_shape}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}
}
\value{
plot file in data folder
}
\description{
Bar plot with dots
Plot by column having WT and KO next on the same bar and all the dots
}
\examples{

table \%>\%
 barplot_dots(x_value = "marker") +
 facet_wrap( ~ treatment)+
 stat_compare_means(
   method = "t.test",
   aes(label = ifelse(p < 0.05, p.signif, round(as.numeric(..p.format..), 2))),
   label.x = 1.5,
   size = 3
 )


}
