% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFU2tidy_genotype.R
\name{CFU2tidy_genotype}
\alias{CFU2tidy_genotype}
\title{CFU to tidy}
\usage{
CFU2tidy_genotype(
  file = c("^CFU", "$csv"),
  path_file = path_raw,
  animalario_file = c("^animalario", "$csv"),
  path_mice = path_raw,
  micecode = micecode,
  animalario_sep = ",",
  multiple = "100",
  marker = "CFU",
  cell = "AM",
  time = "0h",
  treatment = "lp moi"
)
}
\arguments{
\item{file}{.csv with CFU counts. Header should be label with Code for mice
column followed by the different dilutions included in the experiment written
written in the format "Dil-dilution". Example: in dil 1/100 (in 2mL) will be 200
Write NA in raw file when can no count}

\item{path_file}{path where file is located. Usually path_output from
path_builder()}

\item{animalario_file}{raw csv downloaded from animalario with mice used in
the experiment}

\item{path_mice}{path where animalario file to obtain the genotypes is
located. usually path_raw from path_builder.}

\item{micecode}{named list with the replacement for the genotypes.
Load from micecode data included in the package.}

\item{animalario_sep}{separator for the animalario csv file. Default to ","}

\item{multiple}{Number to replace when the CFU where incontable. Represented by M in the original csv}

\item{marker}{usually CFU (by deffault)}

\item{cell}{Am. Indicate if they are from lung, BAL, ex vivo...}

\item{time}{Time after infection. Default to 0h}

\item{treatment}{Moi of infection}
}
\value{
a tibble with the tidy format of the CFU
}
\description{
Function to use when the dilution is already calculated. Standard output. Can read multiple csv at the same time, but is better to do it separately inicating the different times. When multiple, write \emph{M}. When no colonies, write 0.
}
\examples{
data(micecode)
CFU_tidytable_genotype("CFU.csv",  path_file = path_raw, animalario_file = "animalario.csv",
    gate_pattern = gate_pattern, micecode = micecode)


}
