% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFU_boxplot.R
\name{CFU_boxplot}
\alias{CFU_boxplot}
\title{CFU_boxplot}
\usage{
CFU_boxplot(
  table,
  x_lab = "genotype",
  y_lab = "L.pneumophila CFU",
  title_lab = "",
  y_trans = "identity",
  x_angle = NULL,
  x_hjust = NULL,
  color_values = (ggthemes::tableau_color_pal("Classic Green-Orange 12"))(12)[1:12],
  color_breaks = waiver(),
  color_labels = waiver(),
  path_output,
  w = 10,
  h = 5,
  save_plot = FALSE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table coming form CFU_tidytable_genotype}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{y_trans}{transformation for the y axis ("asn", "atanh", "boxcox",
"date", "exp", "hms", "identity", "log", "log10", "log1p", "log2", "logit",
"modulus", "probability", "probit", "pseudo_log", "reciprocal", "reverse",
"sqrt" and "time")}

\item{x_angle}{angle of the labels of the x-axis. NULL for horizontal,
45 for inclination}

\item{x_hjust}{horizontal justification of the labels of the x-axis}

\item{color_values}{a set of aesthetic values to map data values to.
The values will be matched in order (usually alphabetical).}

\item{color_breaks}{takes the limits as input and returns breaks as output}

\item{color_labels}{takes the breaks as input and returns labels as output}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{boolean indicating if the plot is saved or not. Default to TRUE.}

\item{print_plot}{boolean indicating if the plot is printed or not. Default to FALSE.}
}
\value{
plot file in data folder
}
\description{
CFU_boxplot
}
\examples{
CFU_boxplot(table)


}
