% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFU_tidytable_genotype.R
\name{CFU_tidytable_genotype}
\alias{CFU_tidytable_genotype}
\title{CFU_tidytable_genotype}
\usage{
CFU_tidytable_genotype(
  file = c("^CFU", "$csv"),
  path_file = path_raw,
  animalario_file = c("^animalario", "$csv"),
  path_mice = path_raw,
  micecode = micecode,
  animalario_sep = ","
)
}
\arguments{
\item{file}{.csv with CFU counts. Header should be label with Code for mice
column followed by the different dilutions included in the experiment written
written in the format "Dil-1/dilution"}

\item{path_file}{path where file is located. Usually path_output from
path_builder()}

\item{animalario_file}{raw csv downloaded from animalario with mice used in
the experiment}

\item{path_mice}{path where animalario file to obtain the genotypes is
located. usually path_raw from path_builder.}

\item{micecode}{named list with the replacement for the genotypes.
Load from micecode data included in the package.}

\item{animalario_sep}{separator for the animalario csv file. Default to ","}
}
\value{
a tibble with the tidy format of the CFU
}
\description{
CFU_tidytable_genotype
}
\examples{
data(micecode)
CFU_tidytable_genotype("CFU.csv",  path_file = path_raw, animalario_file = "animalario.csv",
    gate_pattern = gate_pattern, micecode = micecode)


}
