% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_bar.R
\name{composition_bar}
\alias{composition_bar}
\title{Bar plot representing cell composition
Plot stacked columns by strain with error bars}
\usage{
composition_bar(
  table,
  plot_cell = "AMs",
  genotype_levels = c("WT", "KO"),
  strain_levels = c("VHL", "VHL-HIF1a", "VHL-HIF2a", "VHL-HIF1a-HIF2a"),
  origin_diff = "-host|-transplant|-SJ",
  group_diff = "-WT|-KO",
  last_origin = "transplant",
  leyend_position = "top",
  x_lab = "",
  y_lab = "Freq. of BAL AMs (\%)",
  title_lab = "",
  y_trans = "identity",
  y_label = waiver(),
  color_values = hue_pal()(200),
  fill_values = hue_pal()(200),
  plot_stat = TRUE,
  path_output,
  w = 10,
  h = 5,
  save_plot = FALSE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with data coming from the analysis. Columns: genotype, value and experiment (time, mice, treatment, marker, stat, cell)}

\item{plot_cell}{cell to plot. Default to "AMs}

\item{genotype_levels}{vector with all the genotypes all the analysis}

\item{strain_levels}{vector with all the strains to be generated for the analysis}

\item{origin_diff}{text to remove that specify different origin inside one chimera}

\item{group_diff}{text to remove from the genotype column to generate the strain by which the relativation groups will be generated. Default to "-WT|-KO"}

\item{last_origin}{name of origin that makes the 100\% in cumsum}

\item{leyend_position}{legend position. Default to top}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{y_trans}{transformation of the y axis}

\item{y_label}{default to waiver. Could be scientific_format()}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{plot_stat}{Boolean indicating if include the stat. Default to TRUE.}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{Boolean indicating if the plot is saved or not. Default to FALSE.}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}
}
\value{
plot file in data folder
}
\description{
Bar plot representing cell composition
Plot stacked columns by strain with error bars
}
\examples{

plot \%>\%
FC_bar(genotype_levels = VHL_table$genotypes,
color_values = VHL_table$palette_color,
shape_values = VHL_table$palette_shape,
fill_values = VHL_table$palette_fill)


}
