% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_init.R
\name{cover_init}
\alias{cover_init}
\title{Generation of the experiment cover in Rmd}
\usage{
cover_init(experiment_name)
}
\arguments{
\item{experiment_name}{Full name of the experiment. Should have the following
structure: project, "hash", year, experiment number, "underscore",
experiment title}
}
\description{
Creation of the cover page in an Rmd to write the protocol and the analysis.
Should be used directly in the console within the project folder
}
\examples{
cover_init("VHL-2101_experiment")


}
