% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_mean_plot.R
\name{curve_mean_plot}
\alias{curve_mean_plot}
\title{curve_mean_plot
plot a graph curve with the mean value of each genotype by one column.
Not finished, it lacks the conection lines}
\usage{
curve_mean_plot(
  table,
  organ.i = NULL,
  experiment.i = NULL,
  stat.i = NULL,
  time.i = NULL,
  marker.i = NULL,
  cell.i = NULL,
  treatment.i = NULL,
  genotype_levels = c("WT", "KO"),
  genotype_labels = genotype_levels,
  x_value = "cell",
  error_width = 1,
  title_lab = "",
  x_lab = "",
  y_lab = "",
  y_limit = 0,
  y_trans = "identity",
  leyend_position = "right",
  color_values = (scales::hue_pal())(200),
  shape_values = rep(21, 200),
  fill_values = (scales::hue_pal())(200),
  lty_values = rep("solid", 200),
  path_output = NULL,
  w = 10,
  h = 5,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table coming form facs_tidytable}

\item{organ.i}{optional organ selected to plot (specimen in .fcs file)}

\item{experiment.i}{optional experiment to select}

\item{stat.i}{optional statistic selected to plot}

\item{time.i}{optional time selected to plot}

\item{marker.i}{optional marker selected to plot}

\item{cell.i}{optional cell selected to plot}

\item{treatment.i}{optional treatment selected to plot}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{genotype_labels}{name to be display in the legend. In markdown/html format.}

\item{x_value}{value to plot in x axis. Default to cell}

\item{error_width}{width of the error bar}

\item{title_lab}{title label}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{y_limit}{y-axis limit}

\item{y_trans}{y-axis transformation}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{shape_values}{shape to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_shape}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{lty_values}{line structure. to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_lty}

\item{path_output}{full name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{print_plot}{boolean indicating if the plot is printed or not. Default to TRUE.}

\item{legend_position}{Legend position. Default to top. Could also be right, left or bottom}
}
\description{
curve_mean_plot
plot a graph curve with the mean value of each genotype by one column.
Not finished, it lacks the conection lines
}
