% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisa_tidytable.R
\name{elisa_tidytable}
\alias{elisa_tidytable}
\title{Make tidytable from ELISA
Function to calculate the concentrations of ELISA experiments giving two files as input, one coming from the spectrometer with the raw OD data and other with a template of the plate used. In the template is expected to be at least 1 point of blank_0 that will be subtracted to all values and two or more values from a standard curve labelled as standard_0, standard_10 ... For samples well the format is name_concentration as proportion of samples (for example a sample diluted 1/2 would be 0.5)}
\usage{
elisa_tidytable(
  data_file = "elisa.txt",
  template_file = "plate.csv",
  path_file = path_raw,
  sep_data = "\\t",
  sep_template = ",",
  model = "log-log",
  max_value = Inf,
  plot_curve = TRUE
)
}
\arguments{
\item{data_file}{txt file with the OD data.}

\item{template_file}{csv file with the template of the experiment. Format will be sample-type_concentration with standard and blank as required keyword}

\item{path_file}{path where data_file and template_file are located. Usually path_raw}

\item{sep_data}{separator character from data_file. Frequently tab}

\item{sep_template}{separator character from template_file. Frequently sep}

\item{model}{model to fit the analysis: either linear ("lm") or 4-parameter log-logistic ("log-log")}

\item{max_value}{to restrict the standard curve to a maximal value}

\item{plot_curve}{print the fitted plot}
}
\description{
Make tidytable from ELISA
Function to calculate the concentrations of ELISA experiments giving two files as input, one coming from the spectrometer with the raw OD data and other with a template of the plate used. In the template is expected to be at least 1 point of blank_0 that will be subtracted to all values and two or more values from a standard curve labelled as standard_0, standard_10 ... For samples well the format is name_concentration as proportion of samples (for example a sample diluted 1/2 would be 0.5)
}
\examples{
elisa_tidytable(
  data_file = "elisa-IFNg.txt",
  template_file = "IFNg-plate.csv",
  path_file = path_raw,
  model = "lm",
  max_value = 50,
  plot_curve = FALSE
)

}
