% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_boxplot.R
\name{facs_boxplot}
\alias{facs_boxplot}
\title{Boxplot creation for cytometry data}
\usage{
facs_boxplot(
  table = "",
  organ.i = NULL,
  stat.i = NULL,
  time.i = NULL,
  marker.i = NULL,
  cell.i = NULL,
  treatment.i = NULL,
  x_value = "cell",
  title.i = "",
  x_lab = "",
  y_lab = "",
  y_limit = 0,
  x_angle = 45,
  x_hjust = 1,
  color_values = (ggthemes::tableau_color_pal("Classic Green-Orange 12"))(12)[1:12],
  color_breaks = waiver(),
  color_labels = waiver(),
  path_output = NULL,
  w = 10,
  h = 5,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table coming form facs_tidytable}

\item{organ.i}{optional organ selected to plot (specimen in .fcs file)}

\item{stat.i}{optional statistic selected to plot}

\item{time.i}{optional time selected to plot}

\item{marker.i}{optional marker selected to plot}

\item{treatment.i}{optional treatment selected to plot}

\item{title.i}{title of the plot}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{y_limit}{inferior limit for y-axis}

\item{x_angle}{angle of the labels of the x-axis. NULL for horizontal,
45 for inclination}

\item{x_hjust}{horizontal justification of the labels of the x-axis}

\item{color_values}{a set of aesthetic values to map data values to.
The values will be matched in order (usually alphabetical).}

\item{color_breaks}{takes the limits as input and returns breaks as output}

\item{color_labels}{takes the breaks as input and returns labels as output}

\item{path_output}{Optional. Full file name desired (e.g. here(path_output, "plot.pdf"))}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{print_plot}{boolean indicating if the plot is printed or not. Default to FALSE.}
}
\value{
plot file in data folder
}
\description{
Boxplot generation for data created with facs_tidytable.
Design to work inside an apply function with all the possible combinations of
parameters in order to generate multiple plots (see examples).
}
\examples{
comb <- as_tibble(unique(paste(table$organ, table$stat, table$marker,
                               sep = "_-_"))) \%>\%
        separate(value, into = c("organ", "stat", "marker"), sep = "_-_") \%>\%
        mutate(output = here(path_output,
                             paste0(organ, "_", stat, "_", marker, ".png")),
               y_lab = paste0(marker, " (", stat, ")")) \%>\% as.data.frame()

apply(comb, 1, function(x) facs_boxplot(table, organ.i = x[1], stat.i = x[2],
  marker.i = x[3], path_output = x[4], y_lab = x[5],
  title.i = x[1]))
}
