% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_tidytable.R
\name{facs_tidytable}
\alias{facs_tidytable}
\title{Prepare the data in a tidy format from the data obtained in Flowjo}
\usage{
facs_tidytable(file, animalario_file, gate_pattern, micecode)
}
\arguments{
\item{file}{.xls generated from Flowjo with cell percentages and fluorescent
intensities}

\item{animalario_file}{raw csv downloaded from animalario with mice used in
the experiment}

\item{gate_pattern}{named list with the replacements desired for the gates.
Load from gate_pattern data included in the package.
Common ones are: c("Lymphocytes/Single Cells/Single Cells/CD452/" = "",
"Freq. of Parent" = "Freq.", "Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\)" = "")}

\item{micecode}{named list with the replacement for the genotypes.
Load from micecode data included in the package.}
}
\value{
a tibble with the tidy format
}
\description{
Generation of a tidytable from the .xls generated from Flowjo.
It's important to have the tubes correctly labelled: specimen should have a
descriptive name without using "_" and each tube should be named using ONLY
the full name of the mice.
}
\examples{
data(gate_pattern)
data(micecode)
facs_tidytable("table.xls", "animalario.csv", gate_pattern = gate_pattern,
    micecode = micecode)


}
