% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_tidytable_genotype.R
\name{facs_tidytable_genotype}
\alias{facs_tidytable_genotype}
\title{Prepare the data in a tidy format from the data obtained in Flowjo}
\usage{
facs_tidytable_genotype(
  file = c("^Table", "$csv"),
  path_file = path_output,
  animalario_file = c("^animalario", "$csv"),
  gate_pattern,
  path_mice = path_raw,
  micecode,
  animalario_sep = ","
)
}
\arguments{
\item{file}{.xls generated from Flowjo with cell percentages and fluorescent
intensities}

\item{path_file}{path where file is located. Usually path_output from
path_builder()}

\item{animalario_file}{raw csv downloaded from animalario with mice used in
the experiment}

\item{gate_pattern}{named list with the replacements desired for the gates.
Load from gate_pattern data included in the package.
Common ones are: c("Freq. of Parent" = "Freq.",
"Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\)" = "")}

\item{path_mice}{path where animalario file to obtain the genotypes is
located. usually path_raw from path_builder.}

\item{micecode}{named list with the replacement for the genotypes.
Load from micecode data included in the package.}

\item{animalario_sep}{separator for the animalario csv file. Default to ","}
}
\value{
a tibble with the tidy format
}
\description{
Generation of a tidytable from the .xls generated from Flowjo.
It's important to have the tubes correctly labelled: specimen should have a
descriptive name without using "_" and each tube should be named using ONLY
the full name of the mice.
}
\examples{
data(gate_pattern)
data(micecode)
facs_tidytable_genotype("table.xls", path_data, "animalario.csv",
    gate_pattern = gate_pattern, micecode = micecode)


}
