% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_tree.R
\name{facs_tree}
\alias{facs_tree}
\title{Generate a gatting tree from a FlowJo Analysis}
\usage{
facs_tree(file_name, path = here::here())
}
\arguments{
\item{file_name}{Name of the .wps file.}

\item{path}{Path where the .wps file locates.}
}
\value{

}
\description{
Plot showing the gatting hierarchy
}
\examples{
facs_tree("C:/Users/elena/Desktop/working on", "analysis.wps")


}
