% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_ttest.R
\name{facs_ttest}
\alias{facs_ttest}
\title{t-test analysis for facs data}
\usage{
facs_ttest(table, file1 = "t-test.csv", file2 = "significant-t-test.csv")
}
\arguments{
\item{table}{name of the table to be analyzed}

\item{file1}{name of the file containing all the analysis}

\item{file2}{name of the file containing only the significant analysis}
}
\value{
print in the screen the significant values

significant_values data table with the significant samples and their
p-value

t-test.csv file with the whole analysis in result folder

significant-t-test.csv file with the samples that show a p-value
lower than 0.05 in result folder
}
\description{
t-test analysis for data coming in a tidy format from facs_tidytable
function.
}
\examples{
facs_ttest(table)
}
