% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_pattern.R
\docType{data}
\name{gate_pattern}
\alias{gate_pattern}
\title{List of replacements for facs analysis}
\format{
Named list with the desired replacements
}
}
\usage{
gate_pattern
}
\description{
Named list containing common replacement for facs analysis
}
\keyword{data}
