% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_paired_violin.R
\name{genotype_paired_violin}
\alias{genotype_paired_violin}
\title{Paired violin plot by genotype}
\usage{
genotype_paired_violin(
  table,
  genotype_levels = c("WT", "KO"),
  y_value = value,
  x_lab = "",
  y_lab = "",
  title_lab = "",
  color_values = hue_pal()(200),
  shape_values = rep(21, 200),
  fill_values = hue_pal()(200),
  path_output,
  w = 10,
  h = 5,
  save_plot = FALSE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with at least the following columns: genotype, value and experiment}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{shape_values}{shape to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_shape}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{Boolean indicating if the plot is saved or not. Default to FALSE.}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}
}
\value{
plot file in data folder
}
\description{
Paired violin plot by genotype
}
\examples{
genotype_violin(genotype_levels = VHL_table$genotypes,
color_values = VHL_table$palette_color,
shape_values = VHL_table$palette_shape,
fill_values = VHL_table$palette_fill)

}
