% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_violin.R
\name{genotype_violin}
\alias{genotype_violin}
\title{Violin plot by genotype}
\usage{
genotype_violin(
  table,
  organ.i = NULL,
  experiment.i = NULL,
  stat.i = NULL,
  time.i = NULL,
  marker.i = NULL,
  cell.i = NULL,
  treatment.i = NULL,
  genotype_levels = c("WT", "KO"),
  genotype_labels = genotype_levels,
  jitter_width = 0.3,
  x_value = "genotype",
  x_lab = "",
  y_lab = "",
  title_lab = "",
  y_trans = "identity",
  y_limit = 0,
  x_angle = 90,
  x_hjust = 0.5,
  x_labels = waiver(),
  leyend_position = "top",
  color_values = colorRamps::primary.colors(),
  shape_values = rep(21, 200),
  fill_values = color_values,
  path_output = NULL,
  w = 10,
  h = 5,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with time, mice, genotype, treatment, marker, stat, value, experiment and cell columns}

\item{organ.i}{optional organ selected to plot (specimen in .fcs file)}

\item{stat.i}{optional statistic selected to plot}

\item{time.i}{optional time selected to plot}

\item{marker.i}{optional marker selected to plot}

\item{cell.i}{optional cell selected to plot}

\item{treatment.i}{optional treatment selected to plot}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{genotype_labels}{name to be display in the legend. In markdown/html format.}

\item{jitter_width}{width of the points}

\item{x_value}{column name to plot}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{y_trans}{transformation of the y axis}

\item{y_limit}{inferior limit for y-axis}

\item{x_angle}{Angle to display the x axis}

\item{x_hjust}{Justificacion of the x axis}

\item{leyend_position}{Legend position. Default to top. Could also be right, left or bottom}

\item{color_values}{color to be ploted. Same number as levels have genotype.}

\item{shape_values}{shape to be ploted. Same number as levels have genotype.}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype.}

\item{path_output}{Optional. Full file name desired (e.g. here(path_output, "plot.pdf"))}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}

\item{y_label}{default to waiver. Could be scientific_format()}
}
\value{
plot file in data folder
}
\description{
Violin for tidy standard data by genotype
}
\examples{
genotype_violin(genotype_levels = VHL_table$genotypes,
color_values = VHL_table$palette_color,
shape_values = VHL_table$palette_shape,
fill_values = VHL_table$palette_fill)

}
