% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_genotype}
\alias{get_genotype}
\title{Generate an object matching miceID with their genotype}
\usage{
get_genotype(file_name = c("^animalario", "$csv"), micecode)
}
\arguments{
\item{file_name}{Input file}

\item{micecode}{Named chr list containing the replacement chr for the
genotype. BBV, CCT, DCX and DCW strains can be loaded
with \code{data(micecode)}}
}
\value{
data.table with two column: one for the miceID and the other with
their genotype
}
\description{
Extract mice number and genotype from multiple files
animalario*.csv pattern as default download
micecode as union from Mote with Genotipado -> stored in data
}
\examples{
data(micecode)
get_genotype("Animalario-VHL2101.csv", micecode)

}
