% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphpad_column.R
\name{graphpad_column}
\alias{graphpad_column}
\title{graphpad_column}
\usage{
graphpad_column(
  name_graphpad = "pzfx",
  sheet_n = 1,
  to_pivot = c(1, 2),
  folder = "",
  output_name = "table"
)
}
\arguments{
\item{name_graphpad}{name of the file to be open. By default select the files with pzfx extension}

\item{sheet_n}{number of the sheet to be read}

\item{to_pivot}{columns selected to pivot}

\item{folder}{folder to look in. Partial pathway from here()}

\item{output_name}{name to be written the tidy_file. In folder data}
}
\value{
tidy table
}
\description{
read a pzfx file containing column displayed data and convert it into a tidy table for further representation
}
\examples{
table <- graphpad_column(file)

}
