% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphpad_grouped.R
\name{graphpad_grouped}
\alias{graphpad_grouped}
\title{graphpad_grouped}
\usage{
graphpad_grouped(
  name_graphpad = "pzfx",
  sheet_n = 1,
  folder = "",
  save = FALSE,
  output_name = "table"
)
}
\arguments{
\item{name_graphpad}{name of the file to be open. By default select the files with pzfx extension}

\item{sheet_n}{number of the sheet to be read}

\item{folder}{folder to look in. Partial pathway from here()}

\item{save}{conditional to decide if save the file}

\item{output_name}{name to be written the tidy_file. In folder data}
}
\value{
tidy table
}
\description{
Return a csv file into data folder
}
\examples{
table <- graphpad_grouped(file)

}
