% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{micecode}
\alias{micecode}
\title{Mice genotypes list}
\format{
plain text
}
\source{
\href{https://animalario.cnic.es/}
}
\usage{
data(micecode)
}
\description{
Named list with the genotypes of the mice I am working with
}
\examples{
data(micecode)
get_genotype("Animalario-VHL2101.csv", micecode)
}
\references{
None
}
\keyword{genotypes}
