% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_builder.R
\name{path_builder}
\alias{path_builder}
\title{Initialize the experiment creating the needed pathways}
\usage{
path_builder(experiment_name)
}
\arguments{
\item{experiment_name}{Full name of the experiment}
}
\description{
Generate the pathways to save the files of the projects and create specific
folders in data, raw and output. The pathways are saved in R environment to
be used along the analysis
}
\examples{
path_builder("VHL-2101_experiment")


}
