% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_ttest.R
\name{qPCR_ttest}
\alias{qPCR_ttest}
\title{ttest analysis form qPCR}
\usage{
qPCR_ttest(
  table,
  genotype_levels,
  name,
  excluded_genes = c("Actin", "empty", "H2O")
)
}
\arguments{
\item{table}{table with the information. Should have the columns
"Cq", "Target", "mice", "Mean_Cq", "Cq_SD", "genotype", "Delta", "Expression"}

\item{genotype_levels}{Select 2 levels to perform the t.test}

\item{name}{name of the output files (before -t-test.CSV)}

\item{excluded_genes}{name of the genes to be excluded in the analysis
#know the present genes in each sample to exclude in the t.test analysis the ones that doesn't appear
table_target <- table \%>\%
group_by(Target, genotype) \%>\%
summarise(n=n()) \%>\%
pivot_wider(id_cols=Target, names_from = genotype, values_from = n, values_fill = 0)}
}
\value{

}
\description{
ttest analysis form qPCR
}
\examples{
qPCR_ttest(table, c("VHL-WT  ", "VHL-KO  "), "VHL")

}
