% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_data.R
\name{relative_data}
\alias{relative_data}
\title{Make relative data to genotype
Make relative data to WT samples from each experiment. representing this as 1.}
\usage{
relative_data(df, group_diff = "-WT|-KO", group_control = "WT")
}
\arguments{
\item{df}{input dataframe with at least genotype and value}

\item{group_diff}{text to remove from the genotype column to generate the strain by which the relativation groups will be generated. Default to "-WT|-KO"}

\item{group_control}{text that identify the control group to relative by. Should be included inside group_diff. Default to "WT"}
}
\description{
Make relative data to genotype
Make relative data to WT samples from each experiment. representing this as 1.
}
\examples{
my_data \%>\% relative_data(.) \%>\% genotype_violin()

}
