% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tidy.R
\name{save_tidy}
\alias{save_tidy}
\title{save_tidy}
\usage{
save_tidy(table, path_output = path_output)
}
\arguments{
\item{table}{final raw table}

\item{path_output}{generated with eps::path_builder()}
}
\value{

}
\description{
save_tidy
}
\examples{
save_tidy(table, path_output)

}
