% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_bar.R
\name{strain_bar}
\alias{strain_bar}
\title{Bar plot by strains
Plot by column having WT and KO next on the same bar}
\usage{
strain_bar(
  table,
  genotype_levels = c("WT", "KO"),
  strain_levels = c("VHL", "VHL-HIF1a", "VHL-HIF2a", "VHL-HIF1a-HIF2a"),
  group_diff = "-WT|-KO",
  identity_bar = "dodge",
  x_lab = "",
  y_lab = "",
  title_lab = "",
  y_trans = "identity",
  y_label = waiver(),
  color_values = hue_pal()(200),
  shape_values = rep(21, 200),
  fill_values = hue_pal()(200),
  path_output,
  w = 10,
  h = 5,
  save_plot = FALSE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with data comming from the analysis. Columns: genotype, value and experiment (time, mice, treatment, marker, stat, cell)}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{strain_levels}{ordered levels to plot. Default to VHL groups}

\item{group_diff}{text to remove from the genotype column to generate the strain by which the relativation groups will be generated. Default to "-WT|-KO"}

\item{identity_bar}{Position to plot the bar graph. "identity" to pile, "dodge" to put next to each other. Default to dodge. Depending on the value, the points represented are pilled up by strain or not}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{y_trans}{transformation of the y axis}

\item{y_label}{default to waiver. Could be scientific_format()}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{shape_values}{shape to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_shape}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{Boolean indicating if the plot is saved or not. Default to FALSE.}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}

\item{x_angle}{Angle to display the x axis}

\item{x_hjust}{Justificacion of the x axis}
}
\value{
plot file in data folder
}
\description{
Bar plot by strains
Plot by column having WT and KO next on the same bar
}
\examples{

plot \%>\%
strain_bar(genotype_levels = VHL_table$genotypes,
color_values = VHL_table$palette_color,
shape_values = VHL_table$palette_shape,
fill_values = VHL_table$palette_fill)


}
