% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  table,
  organ.i = NULL,
  experiment.i = NULL,
  stat.i = NULL,
  time.i = NULL,
  marker.i = NULL,
  cell.i = NULL,
  treatment.i = NULL,
  x_value,
  y_value,
  x_lab = "",
  y_lab = "",
  title_lab = "",
  leyend_position = "right",
  color_levels,
  color_values = colorRamps::primary.colors(),
  shape_values = rep(21, 200),
  fill_values = color_values,
  path_output = NULL,
  w = 10,
  h = 5,
  print_plot = FALSE
)
}
\arguments{
\item{table}{table with at least one column with data coding for the color to plot, one for -log10(pvalue) and other with log2FoldChange}

\item{organ.i}{optional organ selected to plot (specimen in .fcs file)}

\item{stat.i}{optional statistic selected to plot}

\item{time.i}{optional time selected to plot}

\item{marker.i}{optional marker selected to plot}

\item{cell.i}{optional cell selected to plot}

\item{treatment.i}{optional treatment selected to plot}

\item{x_value}{column containing log2 fold change data}

\item{y_value}{column containing -log10(pvalue) data}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{leyend_position}{Legend position. Default to right. Could also be top, left or bottom}

\item{color_levels}{column containing the levels to color with}

\item{color_values}{color to be ploted. Same number as levels have genotype.}

\item{shape_values}{shape to be ploted. Same number as levels have genotype.}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype.}

\item{path_output}{Optional. Full file name desired (e.g. here(path_output, "plot.pdf"))}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{print_plot}{Boolean indicating if the plot is printed or not. Default to TRUE.}
}
\description{
Volcano plot
}
\examples{
volcano_plot(
metabotable2,
cell.i = "BM_cDC1",
x_value = "FC2",
y_value = "pvalue10",
color_levels = "compound",
color_values = c("#990F26","#99600F", "#54990F", "#3D0F99","#0F8299","#333333"),
x_lab = "log2FoldChange",
y_lab = "-log10(pvalue)"
)+
  geom_vline(xintercept=c(-0.6, 0.6), col="darkgrey", lty = 2) +
  geom_hline(yintercept=-log10(0.05), col="darkgrey", lty = 2)

}
