% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_csv_read.R
\name{weight_csv_read}
\alias{weight_csv_read}
\title{weight_csv_read
function to read the csv were the mice weight is storage. Mice are
represented in rows whereas days are represented in columns. To use inside
weight_tidytable function to also get the genotype and make ID as factors.}
\usage{
weight_csv_read(
  csv_file,
  path_csv,
  relative = FALSE,
  csv_sep = ",",
  date_format = "\%d.\%b"
)
}
\arguments{
\item{csv_file}{name of the csv file with the introduced weights.
Usually "weight-curve.csv"}

\item{path_csv}{path were csv file is located. Usually path_output}

\item{relative}{bool indicating if the final output will be raw number or
the relative porcentage to day 1}

\item{csv_sep}{sep parameter for read.delim function. Default ","}

\item{date_format}{format parameter of strftime function. Character string.
The default for the format methods is "\%d.\%b" that is for format "01.jul".
Other common format are \%d-\%m}
}
\value{

}
\description{
weight_csv_read
function to read the csv were the mice weight is storage. Mice are
represented in rows whereas days are represented in columns. To use inside
weight_tidytable function to also get the genotype and make ID as factors.
}
\examples{
weight_csv_read("weight-curve.csv", path_output)



}
