% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_individual_plot.R
\name{weight_individual_plot}
\alias{weight_individual_plot}
\title{weight_individual_plot
plot a weight-loss curve with the individual value of each mice by day}
\usage{
weight_individual_plot(
  table,
  title.i = "",
  x_lab = "",
  y_lab = "",
  y_limit = 0,
  color_values = RColorBrewer::brewer.pal(8, "Paired")[7:8],
  color_breaks = waiver(),
  color_labels = waiver(),
  path_output,
  w = 10,
  h = 5,
  save_plot = TRUE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table coming form facs_tidytable}

\item{title.i}{title of the plot}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{y_limit}{inferior limit for y-axis}

\item{color_values}{a set of aesthetic values to map data values to.
The values will be matched in order (usually alphabetical).}

\item{color_breaks}{takes the limits as input and returns breaks as output}

\item{color_labels}{takes the breaks as input and returns labels as output}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{boolean indicating if the plot is saved or not. Default to TRUE.}

\item{print_plot}{boolean indicating if the plot is printed or not. Default to FALSE.}
}
\value{
plot file in data folder
}
\description{
weight_individual_plot
plot a weight-loss curve with the individual value of each mice by day
}
\examples{
weight_individual_plot(table_raw, y_limit = 10,
path_output = here(path_output, "individual_raw.png"))
}
