% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_relative_curve.R
\name{weight_relative_curve}
\alias{weight_relative_curve}
\title{Plot of relative weight by mice}
\usage{
weight_relative_curve(
  table,
  genotype_levels = c("WT", "KO"),
  x_lab = "Time (days)",
  y_lab = "Relative weight",
  title_lab = "",
  x_angle = 0,
  x_hjust = 0.5,
  color_values = hue_pal()(200),
  shape_values = rep(21, 200),
  fill_values = hue_pal()(200),
  lty_values = rep("solid", 200),
  path_output = "plot.png",
  w = 10,
  h = 5,
  save_plot = FALSE,
  print_plot = FALSE
)
}
\arguments{
\item{table}{tidy table with data comming from the analysis. Columns: time, mice, genotype, value, experiment}

\item{genotype_levels}{vector will all the genotypes all the analysis}

\item{x_lab}{X-axis label}

\item{y_lab}{y-axis label}

\item{title_lab}{title label}

\item{x_angle}{Angle to display the x axis}

\item{x_hjust}{Justificacion of the x axis}

\item{color_values}{color to be ploted. Same number as levels have genotype . For VHL paper table$VHL_palette_color}

\item{shape_values}{shape to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_shape}

\item{fill_values}{fill color to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_fill}

\item{lty_values}{line structure. to be ploted. Same number as levels have genotype. For VHL paper table$VHL_palette_lty}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}

\item{save_plot}{boolean indicating if the plot is saved or not. Default to FALSE.}

\item{print_plot}{boolean indicating if the plot is printed or not. Default to TRUE.}
}
\value{
plot file in data folder
}
\description{
Plot of relative weight by mice
}
\examples{
weight_relative_curve(table, title_lab = "GGP Seahorse",
color_values = table$VHL_palette_color,
shape_values = table$VHL_palette_shape,
fill_values = table$VHL_palette_fill,
lty_values = table$VHL_palette_lty)




}
