% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_statistics.R
\name{weight_statistics}
\alias{weight_statistics}
\title{weight_statistics
Function to perform the statistic analysis of the weight-loss curve experiments.}
\usage{
weight_statistics(
  table_tidy,
  path_to_save = path_output,
  file_name = "statistic.txt"
)
}
\arguments{
\item{table_tidy}{input table in tidy format with columns for genotype, day, value and mice.}

\item{path_to_save}{path where the output txt will be saved (path_output from the path_builder function).}

\item{file_name}{Name of the output file. "statistic.txt" by default.}
}
\value{

}
\description{
weight_statistics
Function to perform the statistic analysis of the weight-loss curve experiments.
}
\examples{
weight_statistics(table_raw, path_to_save = path_output)

}
