% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_tidytable.R
\name{weight_tidytable}
\alias{weight_tidytable}
\title{weight_tidytable}
\usage{
weight_tidytable(
  csv_file,
  path_csv,
  animalario_file,
  path_mice,
  micecode,
  mice_genotype,
  relative = FALSE,
  csv_sep = ",",
  date_format = "\%d.\%b"
)
}
\arguments{
\item{csv_file}{name of the csv file with the introduced weights.
Usually "weight-curve.csv"}

\item{path_csv}{path were csv file is located. Usually path_output}

\item{animalario_file}{raw csv downloaded from animalario with mice used in
the experiment}

\item{path_mice}{path where animalario file to obtain the genotypes is
located. usually path_raw from path_builder.}

\item{micecode}{Named chr list containing the replacement chr for the
genotype. BBV, CCT, DCX and DCW strains can be loaded
with \code{data(micecode)}}

\item{mice_genotype}{ordered list containing the levels of the mice to be
included. Can be obtain from micecode data (followed by /t).
For example c("VHL-HIF2a-WT  ", "VHL-HIF2a-KO  ")}

\item{relative}{bool indicating if the final output will be raw number or
the relative porcentage to day 1}

\item{csv_sep}{sep parameter for read.delim function. Default ","}

\item{date_format}{format parameter of strftime function. Character string.
The default for the format methods is "\%d.\%b" that is for format "01.jul".
Other common format are \%d-\%m}
}
\value{

}
\description{
weight_tidytable
}
\examples{
data(micecode)
weight_tidytable("weight-curve.csv", path_output,
"Animalario-VHL2101.csv", micecode)

}
