% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_boxplot.R
\name{facs_boxplot}
\alias{facs_boxplot}
\title{Boxplot creation for cytometry data}
\usage{
facs_boxplot(
  table,
  organ.i,
  stat.i,
  marker.i,
  title.i = "",
  x_lab = "",
  y_lab = "",
  y_limit = 0,
  color_values = RColorBrewer::brewer.pal(8, "Paired")[7:8],
  color_breaks = waiver(),
  color_labels = waiver(),
  path_output,
  w = 10,
  h = 5
)
}
\arguments{
\item{table}{tidy table coming form facs_tidytable}

\item{organ.i}{organ selected to plot (specimen in .fcs file)}

\item{stat.i}{statistic selected to plot}

\item{marker.i}{marker selected to plot}

\item{title.i}{title of the plot}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{y_limit}{inferior limit for y-axis}

\item{color_values}{a set of aesthetic values to map data values to.
The values will be matched in order (usually alphabetical).}

\item{color_breaks}{takes the limits as input and returns breaks as output}

\item{color_labels}{takes the breaks as input and returns labels as output}

\item{path_output}{ful name of the generated plot including the path
(recommended path_output from path_builder())}

\item{w}{width of the output plot}

\item{h}{high of the output plot}
}
\value{
plot file in data folder
}
\description{
Boxplot generation for data created with facs_tidytable.
Design to work inside an apply function with all the possible combinations of
parameters in order to generate multiple plots (see examples).
}
\examples{
comb <- as_tibble(unique(paste(table$organ, table$stat, table$marker,
                               sep = "_-_"))) \%>\%
        separate(value, into = c("organ", "stat", "marker"), sep = "_-_") \%>\%
        mutate(output = here(path_output,
                             paste0(organ, "_", stat, "_", marker, ".png")),
               y_lab = paste0(marker, " (", stat, ")")) \%>\% as.data.frame()

apply(comb, 1, function(x) facs_boxplot(table, organ.i = x[1], stat.i = x[2],
  marker.i = x[3], path_output = x[4], y_lab = x[5],
  title.i = x[1]))
}
