% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_tidytable.R, R/functions.R
\name{facs_tidytable}
\alias{facs_tidytable}
\title{Prepare the data in a tidy format from the data obtained in Flowjo}
\usage{
facs_tidytable(file, gate_pattern, gates, marker_pattern)

facs_tidytable(file, gate_pattern, gates, marker_pattern)
}
\arguments{
\item{file}{.xls generated from Flowjo with cell percentages and fluorescent
intensities}

\item{gate_pattern}{named list with the replacements desired for the gates.
gate_pattern = c("Lymphocytes/Single Cells/Single Cells/CD452/" = "",
"Freq. of Parent" = "Freq.", "Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\)" = "")}

\item{gates}{named list with the desired replacement for the remaining gates
after the removal of the gate_pattern}

\item{marker_pattern}{}
}
\value{



}
\description{
Generation of a tidytable from the .xls generated from Flowjo.
It's important to have the tubes correctly labelled: specimen should have a
descriptive name without using "_" and each tube should be named using ONLY
the full name of the mice.

Generation of a tidytable from the .xls generated from Flowjo.
It's important to have the tubes correctly labelled: specimen should have a
descriptive name without using "_" and each tube should be named using ONLY
the full name of the mice.
}
\examples{
facs_tidytable("table.xls", gate_pattern = 
c("Lymphocytes/Single Cells/Single Cells/CD452/" = "", 
"Freq. of Parent" = "Freq.", "Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\\\)" = "", 
"Ly6G, Ly6C subset/AMs" = "AMs", "Ly6G, Ly6C subset/EOs" = "EOs",
"Monocytes" = "Mos", "Neutrophils" = "NOs")

 
facs_tidytable("table.xls", gate_pattern = 
c("Lymphocytes/Single Cells/Single Cells/CD452/" = "", 
"Freq. of Parent" = "Freq.", "Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\\\)" = "", 
"Ly6G, Ly6C subset/AMs" = "AMs", "Ly6G, Ly6C subset/EOs" = "EOs",
"Monocytes" = "Mos", "Neutrophils" = "NOs")

 
}
