% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_tidytable.R
\name{facs_tidytable}
\alias{facs_tidytable}
\title{Prepare the data in a tidy format from the data obtained in Flowjo}
\usage{
facs_tidytable(
  file = c("^Table"),
  path_file = path_output,
  time = "0h",
  gate_pattern
)
}
\arguments{
\item{file}{.xls generated from Flowjo with cell percentages and fluorescent
intensities}

\item{path_file}{path where file is located. Usually path_output from
path_builder()}

\item{gate_pattern}{named list with the replacements desired for the gates.
Load from gate_pattern data included in the package.
Common ones are: c("Freq. of Parent" = "Freq.",
"Freq. of Grandparent" = "Freq.",
"Geometric Mean" = "GMFI", "Median" = "MdFI", "\\)" = "")}
}
\value{
a tibble with the tidy format
}
\description{
Generation of a tidytable from the .xls generated from Flowjo.
It's important to have the tubes correctly labelled: specimen should have a
descriptive name without using "_" and each tube should be named using ONLY
the full name of the mice. Generate standard columns: Time, mice, genotype,
treatment, marker, stat, value, experiment, cell
}
\examples{
data(gate_pattern)
facs_tidytable("table.xls", path_file = path_output,
    gate_pattern = gate_pattern)


}
