% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facs_ttest.R
\name{facs_ttest}
\alias{facs_ttest}
\title{t-test analysis for facs data}
\usage{
facs_ttest(
  table,
  path_output = path_output,
  file1 = "t-test.csv",
  file2 = "significant-t-test.csv"
)
}
\arguments{
\item{table}{name of the table to be analyzed. Filter to make that genotype
column contains only 2 factors}

\item{path_output}{path were the output will be stored}

\item{file1}{name of the file containing all the analysis}

\item{file2}{name of the file containing only the significant analysis}
}
\value{
print in the screen the significant values

significant_values data table with the significant samples and their
p-value

t-test.csv file with the whole analysis in output folder

significant-t-test.csv file with the samples that show a p-value
lower than 0.05 in output folder
}
\description{
t-test analysis for data coming in a tidy format from facs_tidytable
function.
}
\examples{
table1 <- filter(genotype \%in\% c("VHL-HIF1a-KO  ", "VHL-HIF1a-WT  "))
facs_ttest(table1,
           path_output = path_output,
           file1 = "HIF1a-DKO-t.test.csv",
           file2 = "HIF1a-DKO-significant-t-test.csv")
}
