% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotype.R
\name{get_genotype}
\alias{get_genotype}
\title{Generate an object matching miceID with their genotype}
\usage{
get_genotype(
  file_name = c("^animalario", "$csv"),
  path_raw,
  micecode,
  csv_sep = ","
)
}
\arguments{
\item{file_name}{Input file}

\item{path_raw}{path were the file is located. Usually in the experiment
subfolder inside raw}

\item{micecode}{Named chr list containing the replacement chr for the
genotype. BBV, CCT, DCX and DCW strains can be loaded
with \code{data(micecode)}}

\item{csv_sep}{separator for the csv file. Default to ","}
}
\description{
Extract mice number and genotype from multiple files
animalario*.csv pattern as default download in english layout
NOTE: take a look to determine if the first row is indicating sep=";"
and remove in that case.
NOTE: if an error ocurrs remove the accent mark.
micecode as union from Nickname with Genotyping -> stored in data
Return a data.table with two column: one for the miceID and the other with
their genotype
}
\examples{
data(micecode)
get_genotype("Animalario-VHL2101.csv", micecode)

}
